==> Rscripts/9-cellphonedb.ipynb <==
import pathlib

import numpy as np
import pandas as pd
import scanpy as sc
from cellphonedb.src.core.methods import cpdb_statistical_analysis_method
from cellphonedb.utils.db_utils import create_db
from scipy.sparse import csr_matrix

np.random.seed(3407)

# Build CellPhoneDB v4.0.0
cpdb_dir = pathlib.Path("/Users/zhangjl/Documents/Proj/Dev_Retina/outputs/CCC/CellPhoneDB/cellphonedb-data-4.0.0/data")
create_db(str(cpdb_dir))

# Load data
proj_dir = pathlib.Path("/Users/zhangjl/Documents/Proj/Dev_Retina/outputs/CCC/")

sc_count = sc.read_csv(proj_dir / "CellPhoneDB/retina_sc.csv").T
sc_count.X = csr_matrix(sc_count.X)

sc_meta = pd.read_csv(proj_dir / "CellPhoneDB/retina_sc_meta.csv", index_col=0)
sc_meta.index = sc_meta.index.str.replace("-1", ".1")

# Inspect data
sc_meta = sc_meta.loc[sc_count.obs_names, :].reset_index().iloc[:, :2]
sc_meta.columns = ["Cell", "cell_type"]
sc_meta.to_csv(
    proj_dir / "CellPhoneDB/cpdb_meta.csv", index=False
)

sc.pp.normalize_total(sc_count)
sc.pp.log1p(sc_count)
sc_count.write_h5ad(
    proj_dir / "CellPhoneDB/cpdb_count.h5ad"
)

# CellPhoneDB
cpdb_results = cpdb_statistical_analysis_method.call(
    cpdb_file_path=str(proj_dir / "CellPhoneDB/cellphonedb-data-4.0.0/cellphonedb.zip"),
    meta_file_path=str(proj_dir / "CellPhoneDB/cpdb_meta.csv"),
    counts_file_path=str(proj_dir / "CellPhoneDB/cpdb_count.h5ad"),
    counts_data="hgnc_symbol",
    threads=4,
    threshold=0.1,
    output_path=str(proj_dir / "CellPhoneDB")
)
